standardize domain architecture, extract db queries, and isolate external schemas
