+++ src/domains/ai-suggestions/ai-suggestions.types.ts
@@ -1,13 +1,14 @@
 import { z } from '@hono/zod-openapi';
 
-import { AI_SUGGESTIONS_CONSTANTS } from './ai-suggestions.constants';
+import env from '@/env';
 
 export const getAiSuggestionsQuerySchema = z.object({
   projectUnitId: z.coerce.number().int().positive(),
   bibleTextIds: z
     .string()
     .regex(/^\d+(,\d+)*$/, 'Expected comma-separated numeric bible text IDs')
-    .describe('Comma-separated list of bible text IDs'),
+    .describe('Comma-separated list of bible text IDs')
+    .transform((val) => val.split(',').map((id) => Number.parseInt(id.trim(), 10))),
 });
 
 export type GetAiSuggestionsQuery = z.infer<typeof getAiSuggestionsQuerySchema>;
@@ -30,12 +31,7 @@ export const queueNextVersesRequestSchema = z.object({
   bookCode: z.string().min(3).max(3),
   chapterNumber: z.number().int().positive(),
   currentVerse: z.number().int().positive(),
-  lookahead: z
-    .number()
-    .int()
-    .positive()
-    .max(20)
-    .default(AI_SUGGESTIONS_CONSTANTS.DEFAULT_LOOKAHEAD),
+  lookahead: z.number().int().positive().max(20).default(env.AI_DEFAULT_LOOKAHEAD),
 });
 
 export type QueueNextVersesRequest = z.infer<typeof queueNextVersesRequestSchema>;
